Updated tests and restructured code
