+++ src/lib/api.ts
@@ -1,10 +1,16 @@
 import jwt from 'jsonwebtoken'
 import {
+    DockerApiResponse,
+    KubernetesApiResponse,
     ManagementAppGetReadyStudiesResponse,
     TOAGetJobsResponse,
     isManagementAppGetReadyStudiesResponse,
     isTOAGetJobsResponse,
 } from './types'
+import http from 'http'
+import https from 'https'
+import { hasReadPermissions } from './utils'
+import { getKubeAPIServiceAccountToken, getNamespace, initHTTPSTrustStore } from './kube'
 
 // Functions for interacting with the Management App
 const generateManagementAppToken = (): string => {
@@ -106,3 +112,154 @@ export const toaUpdateJobStatus = async (
     console.log(`TOA: Status update for job ${jobId} succeeded!`)
     return { success: true }
 }
+
+/* v8 ignore start */
+export const dockerApiCall = async (
+    method: string,
+    path: string,
+    body?: unknown,
+    ignoreResponse: boolean = false,
+): Promise<DockerApiResponse> => {
+    const protocol = process.env.DOCKER_API_PROTOCOL ?? 'https'
+    const host = process.env.DOCKER_API_HOST ?? 'localhost'
+    const port = process.env.DOCKER_API_PORT ?? 443
+    const apiVersion = process.env.DOCKER_API_VERSION ?? 'v1.48'
+    const socketPath = process.env.DOCKER_SOCKET ?? '/var/run/docker.sock'
+    path = path.startsWith('/') ? path : `/${path}`
+    const url = new URL(`${protocol}://${host}:${port}/${apiVersion}${path}`)
+    console.log(`Connecting to docker Engine API: ${url.toString()}`)
+    const options: {
+        hostname: string
+        port: number | undefined
+        path: string
+        method: string
+        socketPath: string | undefined
+        headers: { [key: string]: string }
+    } = {
+        hostname: url.hostname,
+        port: url.port ? parseInt(url.port, 10) : 443,
+        path: url.pathname + url.search,
+        method: method.toUpperCase(),
+        socketPath: undefined,
+        headers: {
+            'Content-Type': 'application/json',
+            'X-Registry-Auth': process.env.DOCKER_REGISTRY_AUTH ?? '',
+        },
+    }
+    console.log(`Headers: ${JSON.stringify(options.headers)}`)
+    let msg: string = ''
+    const canReadDockerSock = hasReadPermissions(socketPath, (error: Error | null) => {
+        if (error) {
+            msg = `Error Accessing file ${socketPath}. Cause: ${JSON.stringify(error)}`
+        } else {
+            msg = `The Docker socket was found with sufficient permissions at: ${socketPath}`
+        }
+    })
+    if (canReadDockerSock) {
+        options.socketPath = socketPath
+    }
+    console.log(`${msg}`)
+    if (method.toUpperCase() === 'POST' && body) {
+        console.log(`Sending POST request to Docker API with body: ${JSON.stringify(body)}`)
+        options.headers['Content-Length'] = Buffer.byteLength(JSON.stringify(body)).toString()
+    }
+    return new Promise((resolve, reject) => {
+        const req = (protocol === 'http' ? http : https).request(options, (response) => {
+            let data = ''
+
+            response.on('data', (chunk) => {
+                data += chunk
+            })
+
+            response.on('end', () => {
+                try {
+                    console.log(`Response: ${data}`)
+                    if (data && !ignoreResponse) {
+                        const result: DockerApiResponse = JSON.parse(data)
+                        if ('message' in result) reject(result)
+                        else resolve(result)
+                    }
+                    resolve({})
+                } catch (error: unknown) {
+                    reject(new Error(`Failed to parse JSON: ${JSON.stringify(error)}`))
+                }
+            })
+        })
+
+        req.on('error', (error) => {
+            console.error('Docker API => Error:', error)
+            reject(error)
+        })
+
+        if (method.toUpperCase() === 'POST' && body) {
+            req.write(JSON.stringify(body))
+        }
+
+        req.end()
+    })
+}
+
+export const k8sApiCall = (
+    group: string,
+    path: string,
+    method: string,
+    body?: unknown,
+): Promise<KubernetesApiResponse> => {
+    const namespace = getNamespace()
+    const kubeAPIServer = process.env.K8S_APISERVER || `https://kubernetes.default.svc.cluster.local`
+    const kubeAPIServerURL = `${kubeAPIServer}/apis/${group}/v1/namespaces/${namespace}/${path}`
+    const kubeAPIServerAccountToken = getKubeAPIServiceAccountToken()
+    initHTTPSTrustStore()
+    console.log(`K8s: Making ${method} => ${kubeAPIServerURL}`)
+    const url = new URL(kubeAPIServerURL)
+    const options: {
+        hostname: string
+        port: number | undefined
+        path: string
+        method: string
+        headers: { [key: string]: string }
+    } = {
+        hostname: url.hostname,
+        port: url.port ? parseInt(url.port, 10) : 443,
+        path: url.pathname + url.search,
+        method: method.toUpperCase(),
+        headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${kubeAPIServerAccountToken}`,
+        },
+    }
+
+    if (method.toUpperCase() === 'POST' && body) {
+        options.headers['Content-Length'] = Buffer.byteLength(JSON.stringify(body)).toString()
+    }
+
+    return new Promise((resolve, reject) => {
+        const req = https.request(options, (response) => {
+            let data = ''
+
+            response.on('data', (chunk) => {
+                data += chunk
+            })
+
+            response.on('end', () => {
+                try {
+                    resolve(JSON.parse(data))
+                } catch (error: unknown) {
+                    reject(new Error(`Failed to parse JSON: ${JSON.stringify(error)}`))
+                }
+            })
+        })
+
+        req.on('error', (error) => {
+            console.error('K8s API => Error:', error)
+            reject(error)
+        })
+
+        if (method.toUpperCase() === 'POST' && body) {
+            req.write(JSON.stringify(body))
+        }
+
+        req.end()
+    })
+}
+/* v8n ignore end */
